# perf: WIP TV primitives